+++ packet.h
@@ -134,9 +134,9 @@ typedef struct {
 /************************************************/
 
 uint32_t packet_encode(mqtt_packet_t *pkt, uint8_t *buf);
-void packet_connect(mqtt_packet_t *pkt, mqtt_connect_opt_t *opt, void *payload, int payload_size);
+void packet_connect(mqtt_packet_t *pkt, mqtt_connect_opt_t *opt, void *payload, size_t payload_size);
 void packet_subscribe(mqtt_packet_t *pkt, mqtt_subscribe_opt_t *opt);
-void packet_publish(mqtt_packet_t *pkt, const char *topic, uint8_t opts, void *payload, int payload_size);
+void packet_publish(mqtt_packet_t *pkt, const char *topic, uint8_t opts, void *payload, size_t payload_size);
 void packet_decode(mqtt_packet_t *pkt, uint8_t *buf);
